internal: fixing structure and logic of code
The previous commiter made a wide array of mistakes in his partial
implementation of broccoli internals, such as gobbing struct with
only unexported fields, faulty compression logic and unnecessary
logging, to name a few. This commit introduces preliminary fixes
to the aforementioned mistakes. That said, I cannot subscribe to
the correctness of this code, as it yet lacks the tests.

For the record: I am the mentor of the previous commiter, hence
he is not to blame for the quality of his code. For him, this is a
part of his learning path, and I'm sure he will learn from it.